worked on setting up gpu

+++ src/cpu/helpers.rs
@@ -3,7 +3,7 @@ use crate::memory::{self, Memory, MemoryType};
 use super::{Cpu, Flag, Register};
 
 pub enum Instruction {
-    Ok(u8, u16, u8, &'static str),
+    Ok(u8, u16, u32, &'static str),
     Invalid(u8),
 }
